Major bug fixes for Skelly and Spider boss
Spider boss now has a collision radius for his First phase attack.
also added a wall that gets set to false when the player starts walking to the boss. and a wall that gets turned on when the player is in the boss room so the player cant go back to bob

Skelly's sphere collider ignores the fireball now so the player should be able to hit his capsule collider from the front now