Audio Source Bug Fix
The audio source bug, where the audio source is calling an object that doesn't exist error has been fixed.

Needed to call on update where audio manager is used to call audio clips, this had to be done in order to get the new audio manager of the next scene.