Bug Fix - Gun clips through wall
Added a secondary camera under main camera that only purpose is to render the weapons in weapon holder object. Fixes the issue with weapons clipping through objects.

Added an assignment to the weaponType variable inside the WeaponPickup method inside the playerScript. This would fix the issue with naming the correct weaponType at weapon pickup.