Bug Fix - Player sprint at wall
Player was able to sprint at a wall and would cause a clipping issue, where the player could see through the other side of the wall.

Fixed by adding an if statement to check if the player is sprinting to modify the players collider radius. Now giving the result of the player not clipping through the wall.